requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
Pillow
requests==2.26.0
python-dotenv==0.19.0
Werkzeug==2.0.3
openai
matplotlib
dashscope
PyJWT
